Implement composable SubtopicDialog for medium and expanded screen width
