+++ app/src/main/java/com/example/studyapp/ui/screens/SubtopicScreen.kt
@@ -33,7 +33,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
 import com.example.studyapp.R
 import com.example.studyapp.data.Subtopic
-import com.example.studyapp.ui.components.SubtopicFullScreenDialog
+import com.example.studyapp.ui.components.study.SubtopicFullScreenDialog
 import com.example.studyapp.ui.viewmodels.SubtopicViewModel
 
 @Composable
